Add bootstrap css and javascript to static files
Add bootstrap css and javascript to the static folder so that the files
can be loaded into Django templates. Create a template for the base
page, which loads the bootstrap css and js to any page that extends the
base page. This will ensure that the styling will be used on any page
that is based on the base template.