Create index page
Add a URL pattern that links to a list of URL patterns for the
application. Add a URL to the app URL patterns that directs to the index
view. The index view responds with the rendered index template, which
has placeholders for different gauges that will update based on live
data received.