Configure Django for SSE
Configure Django to use ASGI to utilize Server Sent Events to open an
event stream between the client and server. This will allow data to be
sent to the clients via python scripts that gather live data. Use
django_eventstream to send the messages and handle connections. Use the
django_eventstream javascript to handle connections and receive
messages.