Add gauge test script, update gauge readings
Add a gauge test script. It will loop infinitely and send events with
random rpm and speed values. Execute this background process as a daemon
process. This will allow it to run in the background while the user is
served the HttpResponse and can continue access the Django server. Add a
javascript file that updates the gauge needle angles to the appropriate
values based on the data from the event stream.