Add meter gauge template, css for fuel and o2
Add a double meter gauge template to create a gauge that has two meters
for different variables to monitor. Add a gauge image for o2 and fuel
consumption. Make the meter display with a dynamic width to allow
measurements to be displayed. Add the o2 and fuel consumption meter to
the index page.