Change gauge naming, test meters
Change the variable naming and set ids to a more general gauge name so
that all gauge types can be referenced with the data_stream_listener.js
file. Add random values to the gauge_test script that will be sent to
the client. This will show that the meters are functional.