Add ability to start/stop lora script
Create two new URLs for two new views. One view starts the lora script,
and the other stops the lora script. The lora script checks a database
flag and exits the loop if the flag is True. The view that stops the
script sets this flag to true, finds the thread, and joins it to
gracefully kill the process.