Update README.md
Added updated information and steps that have been realized through testing up to this point. Add information about the udev rule. Add steps to add the Raspberry Pi ip address to the ALLOWED_HOSTS list in django settings. Update the command to run the server to include the server ip address. These steps are necessary to configure the server to allow other devices to visit the trackside monitor gui page.